Tracking locations of infectious people
By only checking for transmission in locations that have infectious people,
we reduce running time substantially (60% for the Escambia County test case).

The output appears to be valid but is not identical starting in the 4th simulated
month.  This suggests that there is an unstable ordering issue that arises, not
necessarily in the _isHot data structure; it could be somewhere else.  All relevant
sets and maps have a comparator specified, so it's not obvious where the issue is.
Alternatively, it's possible that we have avoided some pointless calls to the RNG,
and that's why the output is different.