Bugfix affecting memory
We were at times tallying events that would happen outside of the running time
of the simulation.  Unfortunately, this did not result in seg faults.  Instead,
it only occasionally (dependent of PRNG seed and parameters) caused problems
during garbage collection, resulting in misleading error messages about other
data structures that were not part of the problem.