Much more sophisticated reporting lag model; better date handling
Previous reporting lag was a fixed value; this commit introduces time-varying,
gamma-distributed time lags.  Reporting times are sampled with a dedicated RNG,
allowing for reporting effects to be studied independent of transmission.

In order to support these changes, we needed to pass around a date object (instead of
a simulation day int), and we needed to be able to handle date strings.