Comorbidity support, reporting refactoring and bugfix
Adding support for comorbidities, and outcome probabilities dependent on age
and comorbid status.

Associating each house with a particular hospital, although currently patients
only go to hospital in a generic sense.  Still need to implement hospital
capacities and transmission to/from/between patients.

Adding Location::dumper() function.  Probably should have one for each class.

Changed reporting parameterization to be probability in each stage, given not
reported in a previous stage.  This ends up being important for getting the
time sequence right, and for guaranteeing that the resulting probabilities of
detection by severity are sensible.

Adding icuMortalityFraction (fraction of real-world deaths that occur in ICU),
used to interpret some of the empirical data fed into the model.

Refactored outcome reporting in some important ways.  Previous approach tended
to reprt cases relatively late (during a more severe stage), inappropriately
neglecting the likeliness of early detection.

Fixed a bug related to the assumption that all people live in a HOUSE.  Used to
be true, but now some people living in NURSHINGHOMES, and quite possibly we'll
be doing something similar for hospitals.