Date conversion bugfix & new seasonality support
Conversion of dates to simulation days worked correctly if the date was in the
same year the simulation started in, but not otherwise.  This was not previously
an issue, but it is now that we're simulating longer time series.

I've also added a first pass at seasonality support, which is currently a catch-all
for any daily multiplicative changes in transmissibility.  Currently, seasonality
affects transmission in all settings.

Renamed julianYear to startJulianYear, since there can be others ...