New hazard-based transmission assumption, expanded VE support
Infections now have an associated relative infectiousness, which can be
modulated by vaccination, and in the future can vary between people or over
time.  To accomodate people being non-uniform in their infectiousness, we're
now using the isHot data structure to track which people at each relative
infectiousness are at each location on each day.  This appears to not have
resulted in a major performance hit, and may have actually improved RAM and
run time.  We'll need to do more thorough testing to know for sure.

To accomodate infectiousness multipliers that could in principle produce
transmissbilities that are >1.0, we've switched to a hazard-based approach.

So that users don't have to specify efficacies for every possible number of
vaccine doses (of particular concern if there is regular boosting in the model),
we now just use whatever the efficacies were for the last dose that had
specified efficacies.

We're now supporting vaccine efficacy against death (VEF) and against
infectiousness (VEI).

Switching to c++17 standard to use nicer loops for maps.