First pass at multi-strain support; other changes
Currently only WILDTYPE and B117 are supported, and in a rudimentary way, but
the framework is in place to have much more nuanced strain support.

This means that strain type is passed from infecter to infectee, and that it is
possible to change what strain is being introduced to the population.  We do
not currently support mutation within the model, only introduced mutants.

To make the above more straightforward, we're now passing pointers for infecter
and location of transmission, instead of IDs.  We also needed to be able to
compare Date objects with human-readable dates, so I've added the necessary
comparison operators.

I also cleaned up a bunch of old stuff related to logging in simulator.h.