DIRTY: v1 of automated social contact fitting implemented
The model can now restore to a cache if the fit of the simulation's
cumulative reported cases to empirical cumulative reported cases
is deemed unacceptable. The model will choose an adjustment (positive
or negative) based on a logistic curve tuned for a specific maximum
allowed adjustment and a specific fitting threshold.

Necessary changes:
    - the fitting system needs to be able to adjust baseline
      transmissibility when the social contact parameter is already 0
        - if this is done, the system needs to clear the cache and reset
          to sim_day 0
    - much cleaning needs to be done on the implementation of the system