Somewhat dirty commit
Working on fixing a fairly minor bug, where the best behavior value is not
always chosen.  Also did a bit of renaming/reducing the number of distance metrics
being calculated.  Code compiles and runs, but we've observed that a given behavior
value does not necessarily produce the same distance metric.  Could be a caching
problem.