Bugfix for bug that never manifested
We changed our immunity model to make natural and vaccine immunity work
similarly, but never moved the parameter assignment out of the vaccine
conditional in vac_impact/main.cpp.  This did not affect any of our results,
however, because we've only been doing simulations with vaccination, so it
was always getting set correctly.