Minor bug fixes
Changing the condition to be vaccinated from < numDoses to == dose. This 
will not change preventing people from being vaccinated with more then 
numDoses. However, this will also ensure that people will not recive 
more than the dose they were queued for at the moment.

Moved dose rollover outside of the vaccination loops. This prevents 
unexpected behavior when doses are pooled. Because all bin/dose combos 
relyu on the same pool of doses, if they are moved in the vaccination 
loop, the doses will be unavabile for other combos regardless of if 
there are still doses available.