Fixed problems with reading in files with dos line endings
For reasons that aren't clear to me (since I thought we were able to read
both dos and unix files without issue), we started having problems reading
in files from Ben's ts directory, which (at present, at least) have dos
line endings.  I added a new safeGetline() function to Utility that will
handle \n, \r\n, and \r line endings.

We still aren't not splitting lines correctly if the last element is an
empty string.  Instead of inserting an empty string after the last del-
imeter, we insert nothing, resulting in a row that is one column shorter
than expected.  I tried a different split function that I found on stack
overflow, but it had the same behavior.  I think the problem, again, is
with the built-in getline() function and how it works with delimeters.