Two bugfixes
New age-based prioritization strategy wasn't being set up correctly, so no one
was getting vaccinated.

In instances where more than 1 realization was attempted per process (e.g. using the -n
flag), process would segfault when we attempt to seed the PRNG at the start of the second
realization.  PRNGs were being freed at the end of the first realization, and couldn't be
safely reallocated because of c-style memory management.  Now when those pointers are freed,
they are set to nullptr, so that we can subsequently verify that they were freed.