Turns waning on, boosts transmission in children, fixes omicron bug
These changes simplify some assumptions, fix a bug, and make other
compensatory changes.  Waning is turned on, age-specific innate
susceptibility is turned off, and transmission in kids is roughly
doubled.  In order to allow the PPB model to correct for residuals,
social transmission is also doubled.  To keep overall transmission
in line, the baseline hazard is reduced by ~2/3.  We also simplify/
debug omicron assumptions by changing the infectious period to be
the same as for all other strains.

With this, and the compensatory changes below, we do as good a job
with fitting, and for the first time do an excellent job with age-
specific seroprevalence.

Compensatory changes include reducing infectiousness of alpha and
delta, decreasing severity of delta while increasing HFR, and
increasing infectiousness of omicron.