Updates to state vac files
This commit pulls new parameters, makefile approaches, etc. into the state vac
project from the active vac project.  We're also having locking problems with
sqlite on the supercomputer that are new and unrelated to anything we've done,
as previously working executables no longer work for torque array jobs.

As a result, there's are two new files, main-no_db.cpp and pars.h, that are a
workaround so that we can run simulations without relying on a db.  This is
certainly not ideal, but it may be a useful approach to keep track of.